Move Kube manifests into README.
Since the manifests need to be edited before they're applied to the
cluster I thought it best to list them in the README rather than as
files so it's more obvious that you can't just kubectl apply them.

I also removed the dns-ssl manifest because it was using
tazjin/kubernetes-letsencrypt which is no longer maintained and because
there are so many ways to do SSL and routing within Kubernetes that I
don't have a specific recommendation.