Use latest atlantis-base.
- Upon merging https://github.com/runatlantis/atlantis/pull/346 I
manually built and pushed a new version of runatlantis/atlantis-base and
tagged it as runatlantis/atlantis-base:v2.0. This change uses that new
tag. Before it was using :latest which isn't good if other people see
that and use atlantis-base:latest because we could break it for them.
- I've also put back the fsContext to 1000 because the atlantis user is
still in that group so we need that for kubernetes.