Adds ability to define a source for atlantis.yaml
Atlantis currently looks for the global atlantis.yaml configuration file
in the branch it pulls down from the PR. This PR adds another flag,
`--config-source`, that is passed to the atlantis server, and defines
the source of the atlantis.yaml file. This now supports only Github, but
can be extended to support other sources in the future,(e.g., Bitbucket,
S3, Google Storage, etc.).

To specify a file within a github repository in a given branch you need
to add this new flag in the following format:

`--config-source="https://github.com/<owner>/<repo>/[path]atlantis.yaml?[ref=<branch>]"`

Replace <owner>, <repo>, and <branch> with the github username, github
repository, and branch name respectively. `[Path]`is optional though.

The `--allow-repo-config` parameter is still required for this feature
to work. In the future, `--allow-repo-config` can be deprecated and its
behaviour can be incorporated in the `--config-source` flag.