Fix testdrive output on solarized theme terminals.
Previously we were using a library to colorize our strings during the
testdrive command. We used [white] which caused the output to not appear
on terminals that use white as the background. This change uses the
default terminal output color instead and only uses colorized strings
for warnings, errors and info via yellow, red and green respectively.
Hopefully those colours still show up on most terminals!