Release locks when closing draft PRs
If plan is run manually in a draft PR, the PR may hold locks. If
the draft PR is then closed, without first marking it Ready For
Review, the locks are left in place. To resolve this, we can
process closing draft PRs in the same way as non-draft PRs.