Force delete of old plan files in already locked PRs
There is an edge case in which if the plan files for modules (1),(2) have
been already created for a commit (a) in a PR and then the user understands
that he made a mistake and he didn't want to change module (2) and drops
the terraform code of it in commit (b), then when the user will try to
run the apply command Atlantis will apply the (2) module plan file as well
as the `PendingPlanFinder.Find` finds all the differences between the upstream
branch and what Atlantis contains and filters out the `.tfplan` files.
There is an issue connected to this bug https://github.com/runatlantis/atlantis/issues/1122

In order to mitigate the above bug after some careful consideration it seems
that the best solution is to drop all the plan files in case the PR has been
locked anytime in the past, which translates that it most likely have created
`.tfplan` files.

This commit is not expected to create any side-effects or introduce any
new problems.