Fix building and publishing of Docker images
https://github.com/runatlantis/atlantis/pull/1943 changed the way Docker images
were built in GitHub Actions so multi-architecture Docker images now are built,
but it unfortunately included a bug which means it hasn't worked yet on any
released version of Atlantis.

The problem is the change expected environment variables to be replaced in the
regular YAML code instead of just the commands which are run.

This attempts to fix that by instead using the template language available to
access the same environment variable.
This is based on the example provided in the documentation here:
https://docs.github.com/en/actions/learn-github-actions/environment-variables.