Add retries to GitHub client for GetModifiedFiles
This is a follow on to resolve similar issues to #1019.

In #1131 retries were added to GetPullRequest. And in #1810 a backoff
was included.

However, those only resolve one potential request at the very
beginning of a PR creation. The other request that happens early on
during auto-plan is one to ListFiles to detect the modified files. This
too can sometimes result in a 404 due to async updates on the GitHub
side.