Skip autoplan if a pull request matches skip keywords for GitHub
This is an attempt to partially support for #932.

If an author of the pull request has a confidence of "no changes for
real resources", it would be great if atlantis could skip autoplan.

The initial implementation of this feature as follows:

If a title of pull request contains the following keywords, skip autoplan.

* [skip atlantis]
* [skip ci]
* [atlantis skip]
* [ci skip]

We should always force to invoke plan with explicit comment command.

This feature is currently implemented only for GitHub just because
I'm a user of GitHub, but I expect it's possible to support other VCS
providers.

Note:

Most of general purpose CI/CD platforms support a concept for skip
build. For examples:
https://circleci.com/docs/2.0/skip-build/
https://docs.github.com/en/actions/guides/about-continuous-integration#skipping-workflow-runs

As far as I know, they check all commits included in the pull request,
not a title of the pull request because they need to support triggered
on push event. On the other hand, the current implementation of atlantis
doesn't triggered on push event and doesn't have all commits on open
event. To simplify the implementation, I think checking the title is
reasonable. Of course it's possible to get all commits included in the
pull request dynamically via additional API calls, please let me know if
we should check commit messages instead of the title.

The original feature request said that the 'keyword' could be
configurable, but I don't think most of users including me need such a
flexibility. So the initial implementation embeds keywords in source. If
someone need to be configurable, feel free to open another feature
request.