Use join.slack.com instead of custom heroku app (#2564)
At the top of the README is a "badge" to join the slack channel and uses a join.slack.com address. Further down is another reference to join Slack but links to a custom Heroku app to send an invite to someone. I think the custom Heroku app is legacy link and could be replaced by the join.slack.com url.