pass githubAppCreds to repo fetcher to generate new tokens on every fetch (#408)
We currently generate a github token on startup and pass it to the RepoFetcher which is used to clone repos. Since the tokens are only valid for a set period of time, we should generate a new token on every fetch. So, we change the RepoFetcher to generate new github token on every fetch.