Add check run cache. (#535)
Currently, we pass around the check run id after creating it for the
first time when we get a signal for a new revision. This worked fine
until we started completing the check runs when changing the state to
"Action Required", once this happens, successive updates to pending, do
not reset the state of the check run so we need to create a new one.
It's not feasible for us to be able to predict where we might need to
create this or update this at any given point in addition to mutating
the check run id that's being passed around everywhere.

Instead it makes more sense to create a central check run cache that
handles creation or update depending on if a check run is complete or
not. This means we know have to cache this data and clean it up
ourselves however we clean it up as soon as we update with a completion
status successfully.

**Note: I've had this in the works for a week or so so i might have
missed something, definitely review thoroughly.**

Additionally: As part of this change i flipped reruns to require
confirmation of the plan, this was what we wanted in the first place and
now that we have this fix + notifications enabled we should be good.