[neptune] Ensure tf interrupts only happen on context cancellation. (#537)
Atlantis graceful shutdown has not been working as intended since the
beginning (ie. 2 years ago). On the atlantis side, we wait for
in-progress tf operations to resolve and go and update all the gh status
checks accordingly before shutting down.
However, our tf subprocess actually immediately gets interrupted when a
sigterm is received because of the way os.exec works in go by default.

os exec processes listen to sigterms of their parents and there is a
specific thing you need to do basically to prevent that from happening
and that's by setting:
```
cmd.SysProcAttr = &syscall.SysProcAttr{
                Setpgid: true,
}
```

In fixing this, I unraveled a couple other issues with shutdown:
1. Having a global output channel doesn't work because we can't predict
when to close it. Previously we were just closing it when the worker
returns but this is incorrect since activities can still be running at
this point. The worker doesn't wait for all resources to be cleaned up
or all activities to be complete.
2. We can't just assume all our jobs are complete and close them since
we drop logs in this case. Instead, we need to wait until all operations
have complete before closing and uploading to s3.

Note: I also noticed here that we will drop init logs if we shutdown in
between init and plan. This is incorrect and we need to fix however,
it's not critical atm since init logs aren't that useful anyways.