Bugfix: TestProjectCommandOutputHandler is unstable
Use buffered channels in TestProjectCommandOutputHandler

The output handler registered with
ProjectCommandOutputHandler.Register() needs the channel that
is given to not block writes.  If the channel
blocks writes, the channel is removed from the list
of active receivers on that JobID.

The tests are set up with unbuffered channels, and there
is a race condition where the channel reader does
not hit `for msg := range ch` before the message that
is sent is handled by AsyncProjectCommandOutputHandler.writeLogLine

By making the channel buffered it does not matter
that the reader isn't reading yet.