Add ScheduleToStartTimeout support to detect worker death. (#544)
**To recap, the core problem is:**
When we are deploying to a temporal worker, the worker will drain all
existing requests before shutting down and starting up and no new work
can get scheduled on that worker.
**Why is this a problem?**
Well this means that any users that need to apply their plan, now have
to wait until the worker is back up before they can do so (since we
store the planfile on disk). It’s rare but terraform operations can be
long lived and can therefore block users.
This is unfortunate because we have multiple workers, and new terraform
plans will have no problem being scheduled on other workers that are
available.

**How does this PR solve this issue?**
We are using temporal’s `ScheduleToStartTimeout` which tracks the time
between when work is scheduled and when it actually starts on a worker.
When this timeout is hit, we fail our child workflow with a retryable
error, therefore it gets rescheduled in its entirety on a new available
worker.

**Failure Scenarios**
1. A rolling deploy just keeps rescheduling your workflow from the
beginning on each single worker (like a domino effect) this is a worst
case though and is unlikely.
AI: Monitor retries and alarm when we've hit the limit.  

2. If our fleet of workers, isn’t capable of handling a large load of
work we could hit this timeout a lot and essentially just keep trying to
reschedule HOWEVER, we do have exponential backoff which should help
with that situation and we’d alarm on scheduling latencies so we’d be
able to scale up our stateful set manually to mitigate.

3. If we are hard down for a period of time all in progress workflows
could be rescheduled at once from the beginning again. So we could blow
out our disk once we come back up since we would basically clone all at
once, HOWEVER we have precautions in place to rate limit work being
pulled from the task queue for situations like that so we should be ok.