Per root neptune rollout (#560)
Creates override level for config WorkflowModeType at the project root
level, that way a root can have its own separate mode. This should be
temporary and removed once platform mode is the only mode available. I
made it a project level config instead of a global level override to
allow for ease of rollout/rollback (similar to what it is overriding).

Also updates the autoplan builder logic to mark the apply check run as
successful only if all roots are on platform mode.

Would there be any locking issues with allowing a repo to support both
modes? We already have a platform mode project runner that wraps the
exist project runner and allows for roots in platform mode to bypass
project locks and apply runs, so it should be safe.