Extend clone lock so it covers outside the "are we current" test
There is a race condition here where we test if we are current, and
only then if we are not current we grab the lock. In the meantime,
that information could be stale.

Extend the lock to cover all operations, and unconditionally wait for
the lock. We can't assume anything can be skipped if we have to wait for
the lock.

Signed-Off-By: Andrew Carter <andrew@emailcarter.com>