Return error on InstrumentedClient.MergePull
When the underlying GitHub Client returns an error it gets swallowed
in this wrapper method. Not only that but also the emitted metrics are
wrong, as it counts one error AND success at the same time.

We've found this @grafana when using Atlantis in a massive repository
with many changes per minute, and sometimes Atlantis leave a comment
saying it's automatically merging but then nothing happens. But
checking the logs, we've found the following error message:

    Unable to merge pull, error: merging pull request: PUT
    https://api.github.com/repos/grafana/redacted/pulls/666/merge: 405
    Base branch was modified. Review and try the merge again. []

And because the error is swallowed and `InstrumentedClient.MergePull`
returns `nil`, then `Automerger` fails to leave a comment saying
merging failed.